First working (somewhat) FSMC interface between STM32 and ICE40.
Two registers are defined, at address 0x0 and 0x2 (first and second
16-bit word). The first bit in each address each controls a LED on the
ICE40 board.

Only write so far, no read.

Doesn't work reliably. It works after flashing the STM32, but then a
reset breaks it until next flash, somehow the NOE and NWE pins stop
working after reset or powercycle of the STM32. Must be some wrong
initialisation of the FSMC on the STM32-side or something. To be
fixed...